Add Post Reports API #5719

+++ post_service/src/main/java/org/khasanof/post_service/dto/post_report/PostReportDetailDTO.java
@@ -17,6 +17,6 @@ public class PostReportDetailDTO extends GenericDTO {
     private PostEntity post;
     private LinkedList<ReportEntity> reports;
     private Integer countReports;
-    private Long totalPointReports;
+    private Integer totalPointReports;
     private Instant lastReportTime;
 }
