+++ post_service/src/main/java/org/khasanof/post_service/dto/post_report/PostReportGetDTO.java
@@ -13,6 +13,6 @@ import java.time.Instant;
 public class PostReportGetDTO extends GenericDTO {
     private String reportPostId;
     private Integer countReports;
-    private Long totalPointReports;
+    private Integer totalPointReports;
     private Instant lastReportTime;
 }
